+++ test/support/active_record.rb
@@ -114,6 +114,8 @@ class Admin < ActiveRecord::Base
   def record_key
     "1"*64
   end
+
+  encrypts :email_address, key_table: "users", key_attribute: "email_ciphertext"
 end
 
 class Agent < ActiveRecord::Base